Extract RubyFileChangeHandler
