+++ lib/ruby_lsp/internal.rb
@@ -43,6 +43,7 @@ require "ruby_lsp/erb_document"
 require "ruby_lsp/rbs_document"
 require "ruby_lsp/store"
 require "ruby_lsp/addon"
+require "ruby_lsp/file_change_handlers/ruby"
 
 # Response builders
 require "ruby_lsp/response_builders/response_builder"
